ServerManager: Fixed BattleSendResult
Changed durationSec from float to int
Changed to only send playerUserIds that are in the game

+++ Assets/QuantumUser/View/Battle/Scripts/Game/BattleGameViewController.cs
@@ -532,7 +532,7 @@ namespace Battle.View.Game
             (
                 BattleParameters.GetPlayerSlotUserIDs(f),
                 (int)e.WinningTeam,
-                (float)e.GameLengthSec,
+                (int)e.GameLengthSec,
                 success =>
                 {
                     if (!success) _debugLogger.Error("Sending battle result failed.");
